+++ crates/kira/src/sound/static_sound/sound/test.rs
@@ -391,7 +391,7 @@ fn waits_for_start_time() {
 /// Without this behavior, a sound in this situation would be in limbo
 /// forever with no way to free up its resources.
 #[test]
-fn stops_if_depending_on_missing_clock() {
+fn stops_if_waiting_on_missing_clock() {
 	let (clock_info_provider, clock_id) = {
 		let mut builder = MockClockInfoProviderBuilder::new(1);
 		let clock_id = builder.add(true, 0, 0.0).unwrap();
@@ -428,6 +428,53 @@ fn stops_if_depending_on_missing_clock() {
 	);
 	sound.on_start_processing();
 	assert_eq!(handle.state(), PlaybackState::Stopped);
+}
+
+/// Tests that a `StaticSound` that had its start time set to a clock time and already
+/// started will not stop if the clock stops.
+#[test]
+fn continues_when_clock_stops() {
+	let (clock_info_provider, clock_id) = {
+		let mut builder = MockClockInfoProviderBuilder::new(1);
+		let clock_id = builder.add(true, 0, 0.0).unwrap();
+		(builder.build(), clock_id)
+	};
+
+	let data = StaticSoundData {
+		sample_rate: 1,
+		frames: (1..100).map(|_| Frame::from_mono(1.0)).collect(),
+		settings: StaticSoundSettings::new().start_time(ClockTime {
+			clock: clock_id,
+			ticks: 0,
+			fraction: 0.0,
+		}),
+		slice: None,
+	};
+	let (mut sound, _) = data.split();
+
+	assert_eq!(
+		sound.process(
+			1.0,
+			&clock_info_provider,
+			&MockModulatorValueProviderBuilder::new(0).build(),
+		),
+		Frame::from_mono(1.0),
+	);
+
+	let clock_info_provider = {
+		let mut builder = MockClockInfoProviderBuilder::new(1);
+		builder.add(false, 0, 0.0).unwrap();
+		builder.build()
+	};
+
+	assert_eq!(
+		sound.process(
+			1.0,
+			&clock_info_provider,
+			&MockModulatorValueProviderBuilder::new(0).build(),
+		),
+		Frame::from_mono(1.0),
+	);
 }
 
 /// Tests that a `StaticSound` can be played partially.